Small refactor for folders.
