+++ automator/frontend/src/app/core/model/values.ts
