+++ automator/frontend/src/app/tags/bindings/bindings.component.ts
@@ -24,10 +24,10 @@ import { MatProgressBarModule } from "@angular/material/progress-bar";
 import { MatFormField } from "@angular/material/select";
 import { MatSort, MatSortModule } from "@angular/material/sort";
 import { MatTableDataSource, MatTableModule } from "@angular/material/table";
-import { Resource, TagBinding, TagsController } from "../../../model/models";
+import { Resource, TagBinding, TagsController } from "../../core/model/models";
 import { AvailableTags } from "../available_tags";
 import { combineLatest } from "rxjs";
-import { Service } from "../../../model/Service";
+import { Service } from "../../core/model/Service";
 import { SingleEditComponent } from "./edit/single-edit/single-edit.component";
 import { BulkEditComponent } from "./edit/bulk-add/bulk-edit.component";
 import { BulkRemoveComponent } from "./edit/bulk-remove/bulk-remove.component";
