+++ automator/frontend/src/app/tags/bindings/edit/bulk-add/bulk-edit.component.ts
@@ -27,12 +27,12 @@ import {
 import { MatIconModule } from "@angular/material/icon";
 import { MatProgressBarModule } from "@angular/material/progress-bar";
 import { MatSnackBar, MatSnackBarModule } from "@angular/material/snack-bar";
-import { Service } from "../../../../../model/Service";
+import { Service } from "../../../../core/model/Service";
 import {
   Resource,
   TagBinding,
   TagsController,
-} from "../../../../../model/models";
+} from "../../../../core/model/models";
 import { EditComponent } from "../edit.component";
 
 type EditTagData = {
